move cover lookup to background job
fixes #3
