+++ app/concepts/book/cover/lookup.rb
@@ -1,31 +1,43 @@
+require_dependency 'book/cover/isbn/lookup'
+require_dependency 'book/cover/proxy/default'
 require_dependency 'bookie/operation'
+require_dependency 'cover'
 
 class Book < ApplicationRecord
   module Cover
     class Lookup < Bookie::Operation
-      ENDPOINT = 'http://bigbooksearch.com/query.php'
+      step :find!
+      step self::Nested(ISBN::Lookup)
+      step self::Model(::Cover)
+      step :proxy!
+      step :fetch!
+      step :process!
+      step :save!
 
-      PARAMS = {SearchIndex: 'books', ItemPage: 1}
-
-      step :params!
-      step :request!
-      step :image!
-      step :url!
-
-      def params!(options, isbn:, **)
-        options['params'] = PARAMS.merge(Keywords: isbn)
+      def find!(options, isbn:, **)
+        options['model'] = ::Cover.find_by(isbn: isbn)
+        options['model'].present? ? RailWay.pass_fast! : true
       end
 
-      def request!(options, params:, **)
-        options['response'] = RestClient.get(ENDPOINT, params: params)
+      def proxy!(options, model:, isbn:, **)
+        options['proxy'] = Proxy::Default.new(model, isbn: isbn)
       end
 
-      def image!(options, response:, **)
-        options['image'] = Nokogiri::HTML(response).css('a img').first
+      def fetch!(options, url:, **)
+        options['image'] = Dragonfly.app.fetch_url(url)
       end
 
-      def url!(options, image:, **)
-        options['url'] = image.attributes['src'].try(:value)
+      def process!(options, proxy:, image:, **)
+        proxy.image(image) do |v|
+          v.process!(:large)         { |job| job.thumb('300x').encode('png') }
+          v.process!(:small)         { |job| job.thumb('100x').encode('png') }
+          v.process!(:large_preview) { |job| job.thumb('300x450#').encode('png') }
+          v.process!(:small_preview) { |job| job.thumb('100x150#').encode('png') }
+        end
+      end
+
+      def save!(options, proxy:, **)
+        proxy.save
       end
     end
   end
