--- app/concepts/book/cover/proxy.rb
@@ -1,14 +1,0 @@
-require_dependency 'bookie/proxy'
-require_dependency 'bookie/proxy/attachable'
-
-class Book < ApplicationRecord
-  module Cover
-    class Proxy < Bookie::Proxy
-      include Bookie::Proxy::Attachable
-
-      property :isbn
-
-      attachable :image
-    end
-  end
-end
