+++ app/concepts/book/cover/proxy/default.rb
@@ -1,0 +1,16 @@
+require_dependency 'bookie/proxy'
+require_dependency 'bookie/proxy/attachable'
+
+class Book < ApplicationRecord
+  module Cover
+    module Proxy
+      class Default < Bookie::Proxy
+        include Bookie::Proxy::Attachable
+
+        property :isbn
+
+        attachable :image
+      end
+    end
+  end
+end
