+++ app/concepts/book/create.rb
@@ -1,6 +1,5 @@
 require_dependency 'book'
 require_dependency 'book/contract/create'
-require_dependency 'book/cover/create'
 require_dependency 'book/guard/create'
 require_dependency 'book/proxy/default'
 require_dependency 'bookie/operation'
@@ -13,8 +12,5 @@ class Book < ApplicationRecord
     step self::Proxy::Contract(constant: Contract::Create)
     step self::Contract::Validate(key: :book)
     step self::Contract::Persist()
-
-    # FIXME move to background job
-    success -> (options, model:, **) { Cover::Create.(nil, isbn: model.isbn) }
   end
 end
