+++ app/concepts/bookie/worker.rb
@@ -1,0 +1,5 @@
+module Bookie
+  class Worker
+    include Sidekiq::Worker
+  end
+end
